WTF-400 TravisCI extracted to new config format
