+++ modules/travisci/settings.go
@@ -1,4 +1,4 @@
-package circleci
+package travisci
 
 import (
 	"os"
@@ -11,14 +11,17 @@ type Settings struct {
 	Common *cfg.Common
 
 	apiKey string
+	pro    bool
 }
 
 func NewSettingsFromYAML(ymlConfig *config.Config) *Settings {
-	localConfig, _ := ymlConfig.Get("wtf.mods.circleci")
+	localConfig, _ := ymlConfig.Get("wtf.mods.travisci")
 
 	settings := Settings{
 		Common: cfg.NewCommonSettingsFromYAML(ymlConfig),
-		apiKey: localConfig.UString("apiKey", os.Getenv("WTF_CIRCLE_API_KEY")),
+
+		apiKey: localConfig.UString("apiKey", os.Getenv("WTF_TRAVIS_API_TOKEN")),
+		pro:    localConfig.UBool("pro", false),
 	}
 
 	return &settings
